♻️ StatusEffect システムをリファクタリングして可読性とメンテナンス性を向上
- 長大な configs マップを機能別ファイルに分割
  - core-states.ts: 重要なゲーム状態 (Dead, Doomed, KnockedOut等)
  - battle-effects.ts: 戦闘エフェクト (Fire, Poison, Stunned等)
  - dream-demon-effects.ts: 夢魔専用エフェクト (Paralysis, Arousal等)
- 冗長な isXXX() メソッドを削除 (重要なもの以外は hasEffect() で代替)
- 循環依存を回避するため StatusEffectTypes.ts を作成
- 既存の機能は完全に保持、破壊的変更なし
